pyproject.toml
[project]
name = "quantra-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anyio==4.4.0",
    "asgiref==3.8.1",
    "attrs==24.2.0",
    "cachetools==5.4.0",
    "cattrs==24.1.0",
    "certifi==2024.7.4",
    "chardet>=5.1.0",
    "croniter==3.0.3",
    "dcrypt==0.0.6",
    "django==5.1",
    "django-q2==1.7.2",
    "django-redis==5.4.0",
    "django-timezone-field==7.0",
    "google-api-python-client>=2.169.0",
    "gunicorn>=23.0.0",
    "h11==0.14.0",
    "hiredis==3.0.0",
    "httpcore==1.0.5",
    "httpx==0.27.0",
    "idna==3.7",
    "inflection==0.5.1",
    "jpype1>=1.4.1",
    "numpy<2.0.0",
    "openpyxl>=3.1.5",
    "orjson>=3.10.18",
    "pandas==2.2.2",
    "psutil==6.0.0",
    "psycopg2-binary==2.9.9",
    "pypdf2>=3.0.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "pytz==2024.1",
    "requests==2.31.0",
    "six==1.16.0",
    "sniffio==1.3.1",
    "sqlparse==0.5.1",
    "ta-lib==0.4.32",
    "tabula-py>=2.7.0",
    "tzdata==2024.1",
    "whitenoise==6.7.0",
    "xlrd>=2.0.1",
    "yfinance>=0.2.58",
]
